% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{lineage2sunburst}
\alias{lineage2sunburst}
\title{Lineage 2 sunburst}
\usage{
lineage2sunburst(
  lineage,
  sep = ">",
  return_dataframe = FALSE,
  verbose = FALSE,
  textorientation = c("horizontal", "radial", "tangential"),
  unit = "N"
)
}
\arguments{
\item{lineage}{e.g. 'Bacteria>Escherichia>Escherichia coli' (character). Can generate for microbes using \code{\link[taxizedbextra:taxid2lineage]{taxizedbextra::taxid2lineage()}}}

\item{sep}{what character separates each parent-child in lineage strings (string)}

\item{return_dataframe}{instead of returning plot, return the dataframe used to generate the plot (boolean)}

\item{verbose}{verbose (boolean)}

\item{textorientation}{orientation sunburst labels. One of "horizontal", "radial", "tangential" (string)}

\item{unit}{would does count represent, e.g. 'reads' or 'samples' (string)}
}
\value{
sunburst plot (plotly) or inheritance + count dataframe if return_dataframe is TRUE
}
\description{
Takes lineage strings e.g. 'Bacteria>Escherichia>Escherichia coli' and generates a sunburst plot.
Size of each segment in sunburst is based on lineage frequency.
}
\examples{
\dontrun{
taxids = c(rep(562, each =  50), rep(28901, times = 20))

 #Convert NCBI taxids to lineage strings
 lineage_strings = taxizedbextra::taxid2lineage(taxids)
 lineage2sunburst(lineage_strings, sep = ">")

}
}
