% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{microbial_sunburst}
\alias{microbial_sunburst}
\title{Microbial Sunburst}
\usage{
microbial_sunburst(
  taxids,
  ultimate_ancestor = "microbial\\ncomposition",
  show_ranks = FALSE,
  ranks_to_include = c("strain", "species", "genus", "family"),
  special_taxid_names = c(unclassified = -1)
)
}
\arguments{
\item{taxids}{NCBI taxonomy IDs (numeric)}

\item{ultimate_ancestor}{Text to add to (string)}

\item{show_ranks}{Should ranks of each lineage level be added to (boolean)}

\item{ranks_to_include}{which ranks to include in lineage string (character)}

\item{special_taxid_names}{for taxids <= 0 you can define special names. (named numeric vector where names = names and values = taxids)}
}
\value{
sunburst figure (plotly)
}
\description{
Build sunburst from taxids
}
\examples{
\dontrun{
taxids = c(rep(561, times = 10), rep(1639, times = 20), rep(529731, times = 10))
microbial_sunburst(taxids)
}
}
