% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{sunburst}
\alias{sunburst}
\title{Sunburst Plot}
\usage{
sunburst(data)
}
\arguments{
\item{data}{data.frame with three columns: Label, Parent and Value. See details for more information}
}
\value{
plotly figure
}
\description{
Sunburst Plot
}
\details{
\tabular{lll}{
   \strong{Column} \tab \strong{Description} \tab \strong{Type} \cr
   \strong{Label} \tab All the terms to be represented in the sunburst plot \tab character \cr
   \strong{Parent} \tab Parent of Label \tab character \cr
   \strong{Value} \tab Value - will determine size of segment in plot \tab numeric \cr
}
}
\examples{
data = data.frame(
  Label=c("E. coli","S. enterica", "L. monocytogenes", "Escherichia", "Salmonella", "Listeria"),
  Parent = c("Escherichia", "Salmonella", "Listeria",  "Bacteria", "Bacteria", "Bacteria"),
  Value = c(20, 10, 100,0, 0, 0)
)
sunburst(data)

}
